refactor(functions): 参拝の能力解析ロジックを performance.js に分離
ユニットテスト可能にするため、副作用のない純粋な計算ロジック
(user_performance / user_formatted_performance / compute_performance_increment /
raw_user_data_from_status / latest_activity_created_at / get_level など)を
index.js から performance.js へ抽出し、index.js からは require して使用する。
関数本体は逐語移設で挙動は不変。エミュレータで全関数のロードを確認済み。
